reworked driveToMinerals method
